# talent-os/backend/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "DraftCreate": {
        "description": "Create an outreach draft \u2014 used by external agents (e.g. a Claude\ncloud agent doing outreach drafting) instead of the in-backend\nOpenRouter drafting job (services/scheduler.py draft_outreach).",
        "properties": {
          "ai_model": {
            "default": "claude",
            "title": "Ai Model",
            "type": "string"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "channel": {
            "default": "email",
            "title": "Channel",
            "type": "string"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "language": {
            "default": "nl",
            "title": "Language",
            "type": "string"
          },
          "presented_candidate_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Presented Candidate Id"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "target_email": {
            "title": "Target Email",
            "type": "string"
          },
          "target_id": {
            "title": "Target Id",
            "type": "integer"
          },
          "target_name": {
            "title": "Target Name",
            "type": "string"
          },
          "target_type": {
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "target_type",
          "target_id",
          "target_email",
          "target_name",
          "subject",
          "body"
        ],
        "title": "DraftCreate",
        "type": "object"
      },
      "DraftUpdate": {
        "properties": {
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          }
        },
        "title": "DraftUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Multi-agent recruitment platform. PostgreSQL + asyncpg + FastAPI BackgroundTasks + Next.js.",
    "title": "Talent OS \u2014 Hermes Recruitment Engine",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/outreach/drafts": {
      "get": {
        "description": "List outreach drafts, optionally filtered by status.",
        "operationId": "list_drafts_api_v1_admin_outreach_drafts_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Drafts",
        "tags": [
          "outreach"
        ]
      },
      "post": {
        "description": "Create a new outreach draft with status='draft'. Never sends \u2014\napproval happens via POST /drafts/{id}/approve. Rejects (409) if a\ndraft already exists for the same (target_type, target_id, job_id)\nwith status='draft', mirroring the dedupe in services/scheduler.py's\ndraft_outreach job.",
        "operationId": "create_draft_api_v1_admin_outreach_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Draft",
        "tags": [
          "outreach"
        ]
      }
    },
    "/api/v1/admin/outreach/drafts/{draft_id}": {
      "put": {
        "description": "Edit a draft's subject/body. Only allowed while status='draft'.",
        "operationId": "update_draft_api_v1_admin_outreach_drafts__draft_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Draft",
        "tags": [
          "outreach"
        ]
      }
    },
    "/api/v1/admin/outreach/drafts/{draft_id}/approve": {
      "post": {
        "description": "Approve a draft and send it. This is the ONLY path that sends an\noutreach email \u2014 it always requires an explicit human action.",
        "operationId": "approve_draft_api_v1_admin_outreach_drafts__draft_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve Draft",
        "tags": [
          "outreach"
        ]
      }
    },
    "/api/v1/admin/outreach/drafts/{draft_id}/reject": {
      "post": {
        "description": "Reject a draft \u2014 it will never be sent.",
        "operationId": "reject_draft_api_v1_admin_outreach_drafts__draft_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject Draft",
        "tags": [
          "outreach"
        ]
      }
    },
    "/api/v1/admin/outreach/run/{job_name}": {
      "post": {
        "description": "Manually trigger one of the scheduled pipeline jobs once, for testing.",
        "operationId": "run_job_api_v1_admin_outreach_run__job_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run Job",
        "tags": [
          "outreach"
        ]
      }
    },
...
  }
}